+++ src/paddle.cpp
@@ -1,7 +1,7 @@
 #include "Paddle.hpp"
 #include "raylib.h"
 #include "Utility.hpp"
-Paddle::Paddle(const vec2d& position, int width, int height, int speed) 
+Paddle::Paddle(const vec2d& position, int width, int height,  float speed) 
 	: m_location{ position }, m_width{ width },
 	  m_height { height }, m_speedx{ speed }, 
 	  m_speedy{ speed },m_score{0}
@@ -37,7 +37,7 @@ int Paddle::get_posY()
 void Paddle::reset()
 {
 	m_location.x = 0;
-		m_location.y = utility::get_window_height() / 2 - 60;
+	m_location.y = utility::get_window_height() / static_cast<float>(2) - 60;
 }
 const int Paddle::get_score() const
 {
